+++ package.json
@@ -15,7 +15,8 @@
     "lintfix": "eslint . --fix",
     "postinstall": "yarn build",
     "migrate": "knex migrate:latest --knexfile knexfile.js",
-    "rollback": "knex migrate:rollback --knexfile knexfile.js"
+    "rollback": "knex migrate:rollback --knexfile knexfile.js",
+    "seed": "knex seed:run --knexfile knexfile.js"
   },
   "dependencies": {
     "body-parser": "~1.17.1",
